Creation of variable texts, which contains list of three sentences.
Added name of the project with details about author.
Created list of valid users.
Created two inputs for username and password.
Created if test to check username and password validity.

created input request to chose one of the three sentences.
Created text analyzer using for and if to get data about uppercase, lowercase, title case words, count of words, count of numbers and sum of numbers in the chosen sentence.